Added a Groceries List
Added a groceries list with three items. The items have a checkbox, which when checked also crosses out the item. Added double click functionality to the list item text itself. Created a remove feature with a trash can icon. Created a header and footer. This was all done using React.